conver svgCanvas.js to TypeScript
